Bugfix for using OpenCV4 and ROS Noetic
The default values for 'face_cascade_name' and 'eyes_cascade_name' were set always except if use_opencv3 was true. Thus, I was not able to use this launch-script with OpenCV4.

As a fix I check whether the arguments are set to a specific OpenCV version and use default values only if this is not the case.